create deploy job
