+++ .github/workflows/deploy.yml
@@ -1,0 +1,40 @@
+name: Deploy imitation Pipeline
+run-name: ${{ github.actor }} is testing out GitHub Actions 🚀
+on:
+  pull_request:
+    branches: [master]
+jobs:
+  testing:
+    name: Deploy imitation
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '20'
+
+      - name: Install dependencies
+        run: npm install
+
+      - name: Build project with Webpack
+        run: npm run build:prod
+
+      - name: Run Oxlint
+        run: npm run lint
+
+      - name: Run unit tests
+        run: npm run test:unit
+
+      - name: Upload coverage reports to Codecov
+        uses: codecov/codecov-action@v5
+        env:
+          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+
+      - name: Run e2e tests
+        run: npm run test:e2e
+
+      - name: Simulated deployment to the server
+        run: echo "Deploying application..."
