+++ manifest.json
@@ -3,6 +3,9 @@
 	"description": "Generates certificates in PDF format",
 	"version": "2.0",
 	"manifest_version": 3,
+	"content_security_policy": {
+		"extension_pages": "script-src 'self'; object-src 'self';"
+	},
 	"icons": {
 		"16": "public/icons/16.png",
 		"32": "public/icons/32.png",
