+++ package.json
@@ -6,14 +6,17 @@
 		"@types/node": "^22.10.9",
 		"babel-jest": "^29.7.0",
 		"codecov": "^3.8.2",
+		"copy-webpack-plugin": "^12.0.2",
 		"jest": "^29.7.0",
 		"jest-environment-jsdom": "^29.7.0",
+		"oxlint": "^0.15.7",
 		"webpack": "^5.97.1",
 		"webpack-cli": "^6.0.1",
-		"oxlint": "^0.15.7"
+		"webpack-obfuscator": "^3.5.1"
 	},
 	"scripts": {
-		"build:dev": "webpack --config webpack.config.js",
+		"build:dev": "webpack --config webpack.dev.js",
+		"build:prod": "webpack --config webpack.prod.js",
 		"test:unit": "jest --coverage",
 		"test:e2e": "",
 		"lint": "npx oxlint ."
