+++ popup.html
@@ -5,7 +5,7 @@
 		<link rel="stylesheet" type="text/css" href="popup.css" />
 		<script src="libs/pdf-lib.min.js"></script>
 		<script src="libs/qrcode.min.js"></script>
-		<script type="module" src="dist/main.js"></script>
+		<script type="module" src="dist/dev_build/main.js"></script>
 		<title>Certificate Extension</title>
 	</head>
 	<body>
